PhotoSwipe Initialization
Minor customization ("hack") of the PhotoSwipe (©Dmitry Semenov) source code
to change the default fragment identifier `gid=n&pid=n` to just `n`.

The second change is to leave the lightbox gallery open after exiting
fullscreen, so that you can actually see, copy and share URL.

This was needed partly because µBlock now blocks the CSS class
`pswp__button--share`, but mostly because closing the gallery automatically
after exiting fullscreen was buggly, so all that code is now finally gone
from photoswipe.js (the top one).

The remainder of photoswipe.js, according to the documentation, "if you
don't need to support IE8, code can be simplified dramatically." [1]

Finally, I've manually modified both the _other_ photoswipe.js and its
minified version. The latter was a pain to edit manually, but I also don't
want to try to replicate the way it once was minified. Catch 22.

[1]: https://photoswipe.com/documentation/getting-started.html