Fix share button bug in PhotoSwipe
PhotoSwipe's share button links to the thumbnail instead of the original.
This has now been fixed with the changes courteously indicated. The changes
have been made to the un-minified code, and the minified code has been
removed.

Unfortunately, this means that previously generated Photog! websites will
not be working anymore, because they linked to the minified code. If you're
reading this, please don't commit minified code to Git because this the kind
of trouble you will end up in.